inventory_manager/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-crispy-forms==2.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
